requirements.txt
streamlit
pandas
numpy
scikit-learn
seaborn
plotly
matplotlib
umap-learn
prince
pyarrow